refactor: split AppEventHandlerScope train orders and naval/army listeners (#3878)
Extract train-dialog order merge helpers into a dedicated library part and
split the combined naval/army session subscription part into separate naval and
army listeners per Phase 2 event-family decomposition.

Refs #3878
