+++ app/lib/core/services/app_event_handler/app_event_handler_scope.dart
@@ -51,9 +51,11 @@ import '../debug/debug_command_helpers.dart' show DebugCommandResult;
 
 part 'app_event_handler_scope_dialog_builders.dart';
 part 'app_event_handler_scope_session_helpers.dart';
+part 'app_event_handler_scope_train_orders.dart';
 part 'app_event_handler_scope_session_subscriptions_observe.dart';
 part 'app_event_handler_scope_session_subscriptions_civilian.dart';
-part 'app_event_handler_scope_session_subscriptions_naval_army.dart';
+part 'app_event_handler_scope_session_subscriptions_naval.dart';
+part 'app_event_handler_scope_session_subscriptions_army.dart';
 part 'app_event_handler_scope_session_subscriptions_diplomacy.dart';
 part 'app_event_handler_scope_session_subscriptions_debug.dart';
 part 'app_event_handler_scope_session_subscriptions.dart';
@@ -85,141 +87,6 @@ const String quickBattleResultDialogId = 'quick_battle_result';
 const String turnNewsDialogId = 'turn_news';
 
 final _logEvent = packageLogger('event');
-
-/// Applies a chosen combat mode to the current game session state.
-@visibleForTesting
-Game? applyCombatModeChoiceToGame(Game? currentGame, CombatMode chosenMode) {
-  if (currentGame == null) {
-    return null;
-  }
-  if (currentGame.defaultCombatMode == chosenMode) {
-    return currentGame;
-  }
-  return currentGame.copyWith(defaultCombatMode: chosenMode);
-}
-
-/// Replaces pending train-at-capital civilian [BuildUnitOrder]s for [humanPlayerId];
-/// keeps military, naval, and other build orders. Matches [TrainCiviliansDialog] semantics.
-Orders _mergeTrainCivilianOrdersForPlayer({
-  required Orders current,
-  required Game game,
-  required String humanPlayerId,
-  required List<BuildUnitOrder> newFromDialog,
-}) {
-  Player? player;
-  for (final p in game.players) {
-    if (p.id == humanPlayerId) {
-      player = p;
-      break;
-    }
-  }
-  final capital = player?.capitalProvinceId;
-  final civilianUnitIds = CivilianEconomyCatalog.byId.keys.toSet();
-  final existingList =
-      current.buildUnitOrdersByPlayerId[humanPlayerId] ??
-      const <BuildUnitOrder>[];
-  final kept = <BuildUnitOrder>[];
-  for (final order in existingList) {
-    final isDialogManaged =
-        !order.isMilitary &&
-        civilianUnitIds.contains(order.unitType) &&
-        capital != null &&
-        order.spawnProvinceId == capital;
-    if (isDialogManaged) {
-      continue;
-    }
-    kept.add(order);
-  }
-  return current.copyWith(
-    buildUnitOrdersByPlayerId: {
-      ...current.buildUnitOrdersByPlayerId,
-      humanPlayerId: [...kept, ...newFromDialog],
-    },
-  );
-}
-
-/// Replaces pending train-at-capital military [BuildUnitOrder]s for [humanPlayerId];
-/// keeps civilian, naval, and other build orders. Matches [TrainMilitaryDialog] semantics.
-Orders _mergeTrainMilitaryOrdersForPlayer({
-  required Orders current,
-  required Game game,
-  required String humanPlayerId,
-  required List<BuildUnitOrder> newFromDialog,
-}) {
-  Player? player;
-  for (final p in game.players) {
-    if (p.id == humanPlayerId) {
-      player = p;
-      break;
-    }
-  }
-  final capital = player?.capitalProvinceId;
-  final regimentIds = RegimentEconomyCatalog.byId.keys.toSet();
-  final existingList =
-      current.buildUnitOrdersByPlayerId[humanPlayerId] ??
-      const <BuildUnitOrder>[];
-  final kept = <BuildUnitOrder>[];
-  for (final order in existingList) {
-    final isDialogManaged =
-        order.isMilitary &&
-        regimentIds.contains(order.unitType) &&
-        capital != null &&
-        order.spawnProvinceId == capital;
-    if (isDialogManaged) {
-      continue;
-    }
-    kept.add(order);
-  }
-  return current.copyWith(
-    buildUnitOrdersByPlayerId: {
-      ...current.buildUnitOrdersByPlayerId,
-      humanPlayerId: [...kept, ...newFromDialog],
-    },
-  );
-}
-
-/// Replaces pending train-at-capital naval [BuildUnitOrder]s for [humanPlayerId];
-/// keeps civilian, military, and other build orders. Naval shares the
-/// `isMilitary: false` flag with civilians, so dialog-managed orders are
-/// additionally filtered by [ShipEconomyCatalog] ship ids. Matches
-/// [TrainNavalDialog] semantics.
-Orders _mergeTrainNavalOrdersForPlayer({
-  required Orders current,
-  required Game game,
-  required String humanPlayerId,
-  required List<BuildUnitOrder> newFromDialog,
-}) {
-  Player? player;
-  for (final p in game.players) {
-    if (p.id == humanPlayerId) {
-      player = p;
-      break;
-    }
-  }
-  final capital = player?.capitalProvinceId;
-  final shipIds = ShipEconomyCatalog.byId.keys.toSet();
-  final existingList =
-      current.buildUnitOrdersByPlayerId[humanPlayerId] ??
-      const <BuildUnitOrder>[];
-  final kept = <BuildUnitOrder>[];
-  for (final order in existingList) {
-    final isDialogManaged =
-        !order.isMilitary &&
-        shipIds.contains(order.unitType) &&
-        capital != null &&
-        order.spawnProvinceId == capital;
-    if (isDialogManaged) {
-      continue;
-    }
-    kept.add(order);
-  }
-  return current.copyWith(
-    buildUnitOrdersByPlayerId: {
-      ...current.buildUnitOrdersByPlayerId,
-      humanPlayerId: [...kept, ...newFromDialog],
-    },
-  );
-}
 
 /// Binds [AppEventHandler] to [appNavigatorKey] for the app lifetime.
 /// SPEC/program/app-event-bus.md (handler); SPEC/program/app-ui-wiring.md (dialog registration).
