+++ app/lib/core/services/app_event_handler/app_event_handler_scope_train_orders.dart
@@ -1,0 +1,136 @@
+part of 'app_event_handler_scope.dart';
+
+/// Applies a chosen combat mode to the current game session state.
+@visibleForTesting
+Game? applyCombatModeChoiceToGame(Game? currentGame, CombatMode chosenMode) {
+  if (currentGame == null) {
+    return null;
+  }
+  if (currentGame.defaultCombatMode == chosenMode) {
+    return currentGame;
+  }
+  return currentGame.copyWith(defaultCombatMode: chosenMode);
+}
+
+/// Replaces pending train-at-capital civilian [BuildUnitOrder]s for [humanPlayerId];
+/// keeps military, naval, and other build orders. Matches [TrainCiviliansDialog] semantics.
+Orders _mergeTrainCivilianOrdersForPlayer({
+  required Orders current,
+  required Game game,
+  required String humanPlayerId,
+  required List<BuildUnitOrder> newFromDialog,
+}) {
+  Player? player;
+  for (final p in game.players) {
+    if (p.id == humanPlayerId) {
+      player = p;
+      break;
+    }
+  }
+  final capital = player?.capitalProvinceId;
+  final civilianUnitIds = CivilianEconomyCatalog.byId.keys.toSet();
+  final existingList =
+      current.buildUnitOrdersByPlayerId[humanPlayerId] ??
+      const <BuildUnitOrder>[];
+  final kept = <BuildUnitOrder>[];
+  for (final order in existingList) {
+    final isDialogManaged =
+        !order.isMilitary &&
+        civilianUnitIds.contains(order.unitType) &&
+        capital != null &&
+        order.spawnProvinceId == capital;
+    if (isDialogManaged) {
+      continue;
+    }
+    kept.add(order);
+  }
+  return current.copyWith(
+    buildUnitOrdersByPlayerId: {
+      ...current.buildUnitOrdersByPlayerId,
+      humanPlayerId: [...kept, ...newFromDialog],
+    },
+  );
+}
+
+/// Replaces pending train-at-capital military [BuildUnitOrder]s for [humanPlayerId];
+/// keeps civilian, naval, and other build orders. Matches [TrainMilitaryDialog] semantics.
+Orders _mergeTrainMilitaryOrdersForPlayer({
+  required Orders current,
+  required Game game,
+  required String humanPlayerId,
+  required List<BuildUnitOrder> newFromDialog,
+}) {
+  Player? player;
+  for (final p in game.players) {
+    if (p.id == humanPlayerId) {
+      player = p;
+      break;
+    }
+  }
+  final capital = player?.capitalProvinceId;
+  final regimentIds = RegimentEconomyCatalog.byId.keys.toSet();
+  final existingList =
+      current.buildUnitOrdersByPlayerId[humanPlayerId] ??
+      const <BuildUnitOrder>[];
+  final kept = <BuildUnitOrder>[];
+  for (final order in existingList) {
+    final isDialogManaged =
+        order.isMilitary &&
+        regimentIds.contains(order.unitType) &&
+        capital != null &&
+        order.spawnProvinceId == capital;
+    if (isDialogManaged) {
+      continue;
+    }
+    kept.add(order);
+  }
+  return current.copyWith(
+    buildUnitOrdersByPlayerId: {
+      ...current.buildUnitOrdersByPlayerId,
+      humanPlayerId: [...kept, ...newFromDialog],
+    },
+  );
+}
+
+/// Replaces pending train-at-capital naval [BuildUnitOrder]s for [humanPlayerId];
+/// keeps civilian, military, and other build orders. Naval shares the
+/// `isMilitary: false` flag with civilians, so dialog-managed orders are
+/// additionally filtered by [ShipEconomyCatalog] ship ids. Matches
+/// [TrainNavalDialog] semantics.
+Orders _mergeTrainNavalOrdersForPlayer({
+  required Orders current,
+  required Game game,
+  required String humanPlayerId,
+  required List<BuildUnitOrder> newFromDialog,
+}) {
+  Player? player;
+  for (final p in game.players) {
+    if (p.id == humanPlayerId) {
+      player = p;
+      break;
+    }
+  }
+  final capital = player?.capitalProvinceId;
+  final shipIds = ShipEconomyCatalog.byId.keys.toSet();
+  final existingList =
+      current.buildUnitOrdersByPlayerId[humanPlayerId] ??
+      const <BuildUnitOrder>[];
+  final kept = <BuildUnitOrder>[];
+  for (final order in existingList) {
+    final isDialogManaged =
+        !order.isMilitary &&
+        shipIds.contains(order.unitType) &&
+        capital != null &&
+        order.spawnProvinceId == capital;
+    if (isDialogManaged) {
+      continue;
+    }
+    kept.add(order);
+  }
+  return current.copyWith(
+    buildUnitOrdersByPlayerId: {
+      ...current.buildUnitOrdersByPlayerId,
+      humanPlayerId: [...kept, ...newFromDialog],
+    },
+  );
+}
